+++ Core/ServiceContext.cs
@@ -29,7 +29,7 @@ namespace Greatbone.Core
         {
         }
 
-        public void ReadData(IDataInput i, int proj = 0)
+        public void ReadData(IDataInput i, short proj = 0)
         {
             i.Get(nameof(shard), ref shard);
             i.Get(nameof(addrs), ref addrs);
@@ -39,7 +39,7 @@ namespace Greatbone.Core
             i.Get(nameof(auth), ref auth);
         }
 
-        public void WriteData<R>(IDataOutput<R> o, int proj = 0) where R : IDataOutput<R>
+        public void WriteData<R>(IDataOutput<R> o, short proj = 0) where R : IDataOutput<R>
         {
             o.Put(nameof(shard), shard);
             o.Put(nameof(addrs), addrs);
@@ -68,7 +68,7 @@ namespace Greatbone.Core
 
         public string password;
 
-        public void ReadData(IDataInput i, int proj = 0)
+        public void ReadData(IDataInput i, short proj = 0)
         {
             i.Get(nameof(host), ref host);
             i.Get(nameof(port), ref port);
@@ -77,7 +77,7 @@ namespace Greatbone.Core
             i.Get(nameof(password), ref password);
         }
 
-        public void WriteData<R>(IDataOutput<R> o, int proj = 0) where R : IDataOutput<R>
+        public void WriteData<R>(IDataOutput<R> o, short proj = 0) where R : IDataOutput<R>
         {
             o.Put(nameof(host), host);
             o.Put(nameof(port), port);
@@ -106,7 +106,7 @@ namespace Greatbone.Core
         // The service id that does signon. Can be null if local
         public string svcid;
 
-        public void ReadData(IDataInput i, int proj = 0)
+        public void ReadData(IDataInput i, short proj = 0)
         {
             i.Get(nameof(mask), ref mask);
             i.Get(nameof(pose), ref pose);
@@ -115,7 +115,7 @@ namespace Greatbone.Core
             i.Get(nameof(svcid), ref svcid);
         }
 
-        public void WriteData<R>(IDataOutput<R> o, int proj = 0) where R : IDataOutput<R>
+        public void WriteData<R>(IDataOutput<R> o, short proj = 0) where R : IDataOutput<R>
         {
             o.Put(nameof(mask), mask);
             o.Put(nameof(pose), pose);
