+++ Sample/City.cs
@@ -12,14 +12,14 @@ namespace Greatbone.Sample
         internal string code;
         internal string[] distrs;
 
-        public void ReadData(IDataInput i, int proj = 0)
+        public void ReadData(IDataInput i, short proj = 0)
         {
             i.Get(nameof(name), ref name);
             i.Get(nameof(code), ref code);
             i.Get(nameof(distrs), ref distrs);
         }
 
-        public void WriteData<R>(IDataOutput<R> o, int proj = 0) where R : IDataOutput<R>
+        public void WriteData<R>(IDataOutput<R> o, short proj = 0) where R : IDataOutput<R>
         {
             o.Put(nameof(name), name);
             o.Put(nameof(code), code);
